[UPD] report_alternative_layout
