+++ report_alternative_layout/i18n/report_alternative_layout.pot
@@ -6,8 +6,8 @@ msgid ""
 msgstr ""
 "Project-Id-Version: Odoo Server 16.0\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2025-05-24 05:54+0000\n"
-"PO-Revision-Date: 2025-05-24 05:54+0000\n"
+"POT-Creation-Date: 2025-05-29 10:52+0000\n"
+"PO-Revision-Date: 2025-05-29 10:52+0000\n"
 "Last-Translator: \n"
 "Language-Team: \n"
 "MIME-Version: 1.0\n"
@@ -16,13 +16,29 @@ msgstr ""
 "Plural-Forms: \n"
 
 #. module: report_alternative_layout
-#: model_terms:ir.ui.view,arch_db:report_alternative_layout.report_custom_address_layout
+#: model_terms:ir.ui.view,arch_db:report_alternative_layout.external_layout_standard
+msgid ":<span> </span>"
+msgstr ""
+
+#. module: report_alternative_layout
+#: model_terms:ir.ui.view,arch_db:report_alternative_layout.external_layout_standard
 msgid "<strong>Remit-to:</strong>"
 msgstr ""
 
 #. module: report_alternative_layout
 #: model:ir.model.fields,field_description:report_alternative_layout.field_ir_actions_report__apply_alternative_layout
+#: model:ir.model.fields,field_description:report_alternative_layout.field_report_paperformat__apply_alternative_layout
 msgid "Apply Alternative Layout"
+msgstr ""
+
+#. module: report_alternative_layout
+#: model:ir.model.fields,field_description:report_alternative_layout.field_ir_actions_report__date_field
+msgid "Date Field"
+msgstr ""
+
+#. module: report_alternative_layout
+#: model_terms:ir.ui.view,arch_db:report_alternative_layout.external_layout_standard
+msgid "Document No:"
 msgstr ""
 
 #. module: report_alternative_layout
@@ -32,6 +48,7 @@ msgstr ""
 
 #. module: report_alternative_layout
 #: model:ir.model.fields,help:report_alternative_layout.field_ir_actions_report__apply_alternative_layout
+#: model:ir.model.fields,help:report_alternative_layout.field_report_paperformat__apply_alternative_layout
 msgid ""
 "If selected, the alternative layout will be applied in the printed report."
 msgstr ""
@@ -44,10 +61,16 @@ msgid ""
 msgstr ""
 
 #. module: report_alternative_layout
-#: model:ir.model.fields,help:report_alternative_layout.field_ir_actions_report__show_header_every_page
+#: model:ir.model.fields,help:report_alternative_layout.field_ir_actions_report__show_address_in_header
+#: model:ir.model.fields,help:report_alternative_layout.field_report_paperformat__show_address_in_header
 msgid ""
 "If selected, the report header will be shown on every page of the report "
 "output."
+msgstr ""
+
+#. module: report_alternative_layout
+#: model:ir.model,name:report_alternative_layout.model_report_paperformat
+msgid "Paper Format Config"
 msgstr ""
 
 #. module: report_alternative_layout
@@ -56,13 +79,19 @@ msgid "Report Action"
 msgstr ""
 
 #. module: report_alternative_layout
+#: model:ir.model.fields,field_description:report_alternative_layout.field_ir_actions_report__show_address_in_header
+#: model:ir.model.fields,field_description:report_alternative_layout.field_report_paperformat__show_address_in_header
+msgid "Show Address In Header"
+msgstr ""
+
+#. module: report_alternative_layout
 #: model:ir.model.fields,field_description:report_alternative_layout.field_ir_actions_report__show_commercial_partner
 msgid "Show Commercial Partner"
 msgstr ""
 
 #. module: report_alternative_layout
-#: model:ir.model.fields,field_description:report_alternative_layout.field_ir_actions_report__show_header_every_page
-msgid "Show Header Every Page"
+#: model:ir.model.fields,field_description:report_alternative_layout.field_ir_actions_report__show_document_number
+msgid "Show Document Number"
 msgstr ""
 
 #. module: report_alternative_layout
