+++ report_alternative_layout/report/report_template.xml
@@ -1,6 +1,41 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <odoo>
     <template id="external_layout_standard" inherit_id="web.external_layout_standard">
+        <xpath expr="//div[@class='col-9 text-end']" position="attributes">
+            <attribute name="style">display: none;</attribute>
+        </xpath>
+        <xpath expr="//div[@class='col-9 text-end']" position="after">
+            <div class="col-9 text-end">
+                <t
+                    t-if="company.report_header and company.report_header != '&lt;p&gt;&lt;br&gt;&lt;/p&gt;'"
+                >
+                    <span t-field="company.report_header" name="moto" />
+                </t>
+                <!-- Display the document number in header -->
+                <t t-if="report and report.show_document_number">
+                    <div class="col text-end" t-if="o.name">
+                        <strong>
+                            Document No:
+                            <span t-field="o.name" />
+                        </strong>
+                    </div>
+                </t>
+                <!-- Display the date field in header -->
+                <t t-if="report and report.date_field and o">
+                    <t
+                        t-set="date_value"
+                        t-value="report._get_date_field_value(o, report.date_field)"
+                    />
+                    <t t-if="date_value">
+                        <div class="col text-end">
+                            <t t-esc="report.date_field.field_description" />
+                            :<span>&#xa0;</span>
+                            <span t-esc="date_value" />
+                        </div>
+                    </t>
+                </t>
+            </div>
+        </xpath>
         <!-- Hide company address in the header if alternative layout is selected -->
         <xpath expr="//div[@name='company_address']/ul" position="attributes">
             <attribute
@@ -39,32 +74,26 @@
             >
                 <t t-call="report_alternative_layout.report_custom_address_layout" />
             </t>
-            <t
-                t-set="report"
-                t-value="o.env['ir.actions.report']._get_report_from_name(xmlid)"
-            />
-            <t t-set="remit_to_bank" t-value="report._get_remit_to_bank(o)" />
-            <div class="row">
-                <div class="col-7"></div>
-                <div name="remit_to_bank" t-if="remit_to_bank" class="mt16 col-5">
-                    <strong>Remit-to:</strong>
-                    <ul class="list-unstyled">
-                        <li><span t-field="remit_to_bank.bank_id" /></li>
-                        <li><span t-field="remit_to_bank.acc_number" /></li>
-                        <li>
-                            <span t-field="remit_to_bank.acc_holder_name" />
-                        </li>
-                    </ul>
+            <t t-if='report'>
+                <t t-set="remit_to_bank" t-value="report._get_remit_to_bank(o)" />
+                <div class="row">
+                    <div class="col-7" />
+                    <div name="remit_to_bank" t-if="remit_to_bank" class="mt16 col-5">
+                        <strong>Remit-to:</strong>
+                        <ul class="list-unstyled">
+                            <li><span t-field="remit_to_bank.bank_id" /></li>
+                            <li><span t-field="remit_to_bank.acc_number" /></li>
+                            <li>
+                                <span t-field="remit_to_bank.acc_holder_name" />
+                            </li>
+                        </ul>
+                    </div>
                 </div>
-            </div>                       
+            </t>
         </xpath>
     </template>
     <template id="report_custom_address_layout">
-        <div class="pt-3">
-            <t
-                t-set="report"
-                t-value="o.env['ir.actions.report']._get_report_from_name(xmlid)"
-            />
+        <div class="pt-1">
             <t t-set="partner" t-value="report._get_partner(o.partner_id)" />
             <div class="row" name="address_layout">
                 <div class="col-7" name="partner_address">
@@ -93,5 +122,15 @@
                 </div>
             </div>
         </div>
+    </template>
+    <template id="external_layout" inherit_id="web.external_layout">
+        <xpath expr="//t[@t-if='not company']" position="after">
+            <t t-if="o and o.env">
+                <t
+                    t-set="report"
+                    t-value="o.env['ir.actions.report']._get_report_from_name(xmlid)"
+                />
+            </t>
+        </xpath>
     </template>
 </odoo>
